Make dbus activation sandbox-aware
When we call org.freedesktop.Application.Open to activate
an application and pass file uris, the application may not
be able to see the files due to a flatpak sandbox. In this
case, call the document portal to export the files and pass
the resulting uri's instead of the original ones.